+++ CHANGELOG.rst
@@ -12,7 +12,7 @@ Change Log
 Unreleased
 ----------
 
-*
+* Add proctoring test settings since this had the wrong proctoring version.
 
 [2.13.0] - 2020-06-17
 ---------------------